pyproject.toml
[tool.poetry]
name = "demo-fastapi-ci-cd"
version = "0.1.0"
description = "This project demonstrates how to use FastApi framework to creat apis"
authors = ["Reza Mohammadi <rmm.dev.backend@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.114.1"}
pytest = "^8.3.3"
docker = "^7.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvloop = "^0.20.0"
httptools = "^0.6.1"
passlib = "^1.7.4"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]